+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 unreleased
 ==========
+* fix: Content Expiry Changelist: Removed polymorphic model support due to persistant timeouts and low performance
 
 0.0.5 (2021-11-16)
 ==================